[4860] added LabMapping validation, and deletion on LabItem delete

+++ ch.elexis.core.ui.laboratory/src/ch/elexis/core/ui/laboratory/preferences/LaborPrefs.java
@@ -13,8 +13,10 @@
 
 package ch.elexis.core.ui.laboratory.preferences;
 
+import java.text.MessageFormat;
 import java.util.List;
 
+import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.layout.TableColumnLayout;
 import org.eclipse.jface.preference.PreferencePage;
 import org.eclipse.jface.viewers.ColumnLabelProvider;
@@ -61,6 +63,7 @@ import ch.elexis.core.ui.laboratory.commands.CreateMergeLabItemUi;
 import ch.elexis.core.ui.laboratory.commands.EditLabItemUi;
 import ch.elexis.core.ui.util.SWTHelper;
 import ch.elexis.data.LabItem;
+import ch.elexis.data.LabMapping;
 import ch.elexis.data.LabResult;
 import ch.elexis.data.Query;
 
@@ -357,14 +360,18 @@ public class LaborPrefs extends PreferencePage implements IWorkbenchPreferencePa
 				Object o = sel.getFirstElement();
 				if (o instanceof LabItem) {
 					LabItem li = (LabItem) o;
-					Query<LabResult> qbe = new Query<LabResult>(LabResult.class);
-					qbe.add("ItemID", "=", li.getId()); //$NON-NLS-1$ //$NON-NLS-2$
-					List<LabResult> list = qbe.execute();
-					for (LabResult po : list) {
-						po.delete();
+					if (MessageDialog.openQuestion(getShell(), Messages.LaborPrefs_deleteItem,
+						MessageFormat.format(Messages.LaborPrefs_deleteReallyItem,
+							li.getLabel()))) {
+						if (deleteResults(li)) {
+							deleteMappings(li);
+							li.delete();
+							tableViewer.remove(li);
+						} else {
+							MessageDialog.openWarning(getShell(), Messages.LaborPrefs_deleteItem,
+								Messages.LaborPrefs_deleteFail);
+						}
 					}
-					li.delete();
-					tableViewer.remove(o);
 				}
 			}
 		});
@@ -377,14 +384,18 @@ public class LaborPrefs extends PreferencePage implements IWorkbenchPreferencePa
 					Messages.LaborPrefs_deleteAllExplain)) {
 					Query<LabItem> qbli = new Query<LabItem>(LabItem.class);
 					List<LabItem> items = qbli.execute();
+					boolean success = true;
 					for (LabItem li : items) {
-						Query<LabResult> qbe = new Query<LabResult>(LabResult.class);
-						qbe.add("ItemID", "=", li.getId()); //$NON-NLS-1$ //$NON-NLS-2$
-						List<LabResult> list = qbe.execute();
-						for (LabResult po : list) {
-							po.delete();
+						if (deleteResults(li)) {
+							deleteMappings(li);
+							li.delete();
+						} else {
+							success = false;
 						}
-						li.delete();
+					}
+					if (!success) {
+						MessageDialog.openWarning(getShell(), Messages.LaborPrefs_deleteAllItems,
+							Messages.LaborPrefs_deleteFail);
 					}
 					tableViewer.refresh();
 				}
@@ -392,6 +403,31 @@ public class LaborPrefs extends PreferencePage implements IWorkbenchPreferencePa
 		});
 		if (CoreHub.acl.request(AccessControlDefaults.DELETE_LABITEMS) == false) {
 			bDelAllItems.setEnabled(false);
+		}
+	}
+	
+	private boolean deleteResults(LabItem li){
+		boolean ret = true;
+		Query<LabResult> qbe = new Query<LabResult>(LabResult.class);
+		qbe.add(LabResult.ITEM_ID, "=", li.getId()); //$NON-NLS-1$ //$NON-NLS-2$
+		List<LabResult> list = qbe.execute();
+		for (LabResult po : list) {
+			if (CoreHub.getLocalLockService().acquireLock(po).isOk()) {
+				po.delete();
+				CoreHub.getLocalLockService().releaseLock(po);
+			} else {
+				ret = false;
+			}
+		}
+		return ret;
+	}
+	
+	private void deleteMappings(LabItem li){
+		Query<LabMapping> qbe = new Query<LabMapping>(LabMapping.class);
+		qbe.add(LabMapping.FLD_LABITEMID, "=", li.getId()); //$NON-NLS-1$ //$NON-NLS-2$
+		List<LabMapping> list = qbe.execute();
+		for (LabMapping po : list) {
+			po.delete();
 		}
 	}
